requirements.txt
alembic==1.9.2
aniso8601==9.0.1
attrs==22.2.0
Babel==2.11.0
blinker==1.4
Brotli==1.0.9
cachelib==0.10.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.0
cssselect2==0.7.0
dnspython==2.3.0
email-validator==1.1.3
Faker==13.15.1
Flask==2.0.2
Flask-Charts==1.7
Flask-Dance==5.1.0
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-Minify==0.37
Flask-RBAC==0.5.0
flask-restx==0.5.1
Flask-Seeder==1.2.0
flask-session>=0.3.2,<0.5
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
fonttools==4.38.0
fpdf==1.7.2
future==0.18.3
greenlet==2.0.2
gunicorn==20.1.0
html5lib==1.1
htmlmin==0.1.12
identity>=0.2,<0.3
idna==3.4
itsdangerous==2.1.2
jinja-datatables==0.0.8
Jinja2==3.0.2
jsmin==3.0.1
jsonschema==4.17.3
lesscpy==0.15.1
Mako==1.2.4
MarkupSafe==2.1.2
msal==1.18.0
oauthlib==3.2.2
pdfkit==1.0.0
Pillow==9.4.0
ply==3.11
pycparser==2.21
pydyf==0.5.0
PyJWT==2.6.0
pymssql==2.2.7
PyMySQL==1.0.2
pyOpenSSL==23.0.0
pyphen==0.13.2
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==0.19.2
python-highcharts==0.4.2
pytz==2022.7.1
rcssmin==1.1.1
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
SQLAlchemy==1.4.29
SQLAlchemy-Utils==0.40.0
sqlalchemy-filters==0.12.0
tinycss2==1.2.1
Unidecode==1.3.4
urllib3==1.26.14
URLObject==2.4.3
weasyprint==57.2
webencodings==0.5.1
Werkzeug==2.0.3
WTForms==3.0.0
xxhash==3.2.0
zopfli==0.2.2

zipp==3.8.1
git+https://github.com/azure-samples/ms-identity-python-samples-common
flask_rbac
